Fix file adding via QtC
For a some time QtC disables files adding to the UTILITY targets. But
custom target is a utility target. So, as WA make it target as STATIC
library that excluded from the ALL targets build with forces CXX linker
language.

If you manually run build process for this target it will obviously
fail. So, be careful :-)

Also, added third way to add other_files target without modification of
the original code based on the QtC package manager configuration. Look
into README.md for details.